Modification structure pour générer jar unique pour client et server
